+++ Release/_hijack_root/qml/BasicControls/ToolButton.qml
@@ -2,7 +2,6 @@ import QtQuick 2.3
 import QtQuick.Controls 1.2
 import QtQuick.Controls.Styles 1.2
 import "." as Local
-import "../Singletons"
 import "../_OneMakerMV"
 
 ToolButton {